Add files via upload
This example showed how to use MySQL to make the producer idempotent and deal with panics in a Go application. We demonstrated how to connect to the MySQL server using the database/sql package, configure RequiredAcks and IdempotentProducer, and run many SQL operations simultaneously using goroutines. In order to handle panics and hand control back to the calling programme, the recover function was also utilised in a defer statement.

In a Go application that uses MySQL as the database, this example can be helpful as a jumping-off point for implementing idempotent and robust producer behaviour. The example can be changed to meet your unique needs and specifications.